refactor: KAN-155 repository 패키지 이동
